requirements.txt
streamlit
langchain
python-dotenv
langchain-community
pypdf
sentence-transformers
faiss-cpu
openai